Cleaning the files
